remove composable from everything in prep for removal.  Clean up unused files
